Initial Commit
Our game is called Servo Soccer Smackdown, and it is essentially a table top recreation of a simple soccer style game. Each player has a “kicker” which consists of an arduino board hooked up to an ultrasonic sensor and a servo motor which can be used to “kick” the ping pong ball back and forth. Each player gets goal which is placed behind their “kicker” with a bear figurine acting as a goalie. The players must compete to hit the ping pong ball into the opponent’s goal, and the first player to reach 10 points wins the game.  	 	The data collected by the ultrasonic sensor is converted to in and cm, then these are used to map a variable named angle that controls the angle of the servo motor. The map function is used to map angle to (cm, 0) to start, then sets a range of (0, 179).